--- src/App.css
@@ -1,7 +1,0 @@
-.App {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  min-height: 100dvh;
-}
