+++ src/components/Greetings.tsx
@@ -1,9 +1,8 @@
-import styleclass from './Greeting.module.css'
 const Greetings = () => {
   return (
-    <div className={styleclass.card}>
-      <div className={styleclass.header}>Greetings</div>
-      <div className={styleclass.body}>
+    <div className="w-64 h-auto border-solid border border-black rounded-md flex flex-col justify-around items-center gap-2 p-5">
+      <div className=" text-lg text-red-600">Greetings</div>
+      <div className="text-base ">
         Lorem ipsum dolor sit, amet consectetur Lorem ipsum dolor sit amet consectetur adipisicing elit. Iure
         perspiciatis repellat repudiandae nesciunt ullam nulla officiis error odit qui veniam.
       </div>
