--- src/components/Greetings/Greeting.module.css
@@ -1,23 +1,0 @@
-.card {
-  width: 250px;
-  height: 200px;
-  border: 1px solid;
-  border-radius: 5px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-around;
-  align-items: center;
-
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  text-wrap: wrap;
-  & > .body {
-    font-size: 18px;
-    height: 30px;
-  }
-  & > .header {
-    font-size: 36px;
-    color: brown;
-  }
-}
