--- src/components/MainContent/MainContent.module.css
@@ -1,6 +1,0 @@
-.maincontent {
-  flex: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
