+++ src/components/Maincontent.tsx
@@ -1,0 +1,9 @@
+import Greetings from './Greetings'
+const Maincontent = () => {
+  return (
+    <main className="flex-1 flex items-center justify-center">
+      <Greetings />
+    </main>
+  )
+}
+export default Maincontent
