+++ src/components/Navbar.tsx
@@ -1,0 +1,11 @@
+const Navbar = () => {
+  return (
+    <nav className="h-8 py-5 px-3 w-screen flex sticky top-0 justify-between items-center bg-slate-100">
+      <div>Navbar</div>
+      <div>
+        <button>login</button>
+      </div>
+    </nav>
+  )
+}
+export default Navbar
