--- src/components/Navbar/Navbar.module.css
@@ -1,11 +1,0 @@
-.navbar {
-  height: 10vh;
-  padding: 10px 20px;
-  display: flex;
-  position: sticky;
-  top: 0px;
-  justify-content: space-between;
-  width: 100vw;
-  align-items: center;
-  background-color: aliceblue;
-}
